requirements.txt
-i https://pypi.org/simple
alembic==1.13.1; python_version >= '3.8'
aniso8601==9.0.1
appnope==0.1.4; sys_platform == 'darwin'
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.1.3; python_version >= '3.7'
blinker==1.8.2; python_version >= '3.8'
cachelib==0.13.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
decorator==5.1.1; python_version < '3.11' and python_version >= '3.7'
executing==2.0.1; python_version >= '3.5'
faker==25.1.0; python_version >= '3.8'
flask==3.0.3; python_version >= '3.8'
flask-bcrypt==1.0.1
flask-cors==4.0.1
flask-marshmallow==1.2.1; python_version >= '3.8'
flask-migrate==4.0.7; python_version >= '3.6'
flask-restful==0.3.10
flask-session==0.8.0; python_version >= '3.8'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
importlib-metadata==7.1.0; python_version < '3.10'
importlib-resources==6.4.0; python_version < '3.9'
ipdb==0.13.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ipython==8.12.3; python_version < '3.11' and python_version >= '3.7'
itsdangerous==2.2.0; python_version >= '3.8'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
mako==1.3.3; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.21.2; python_version >= '3.8'
marshmallow-sqlalchemy==1.0.0; python_version >= '3.8'
matplotlib-inline==0.1.7; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
msgspec==0.18.6; python_version >= '3.8'
packaging==24.0; python_version >= '3.7'
parso==0.8.4; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
psycopg2-binary==2.9.9; python_version >= '3.7'
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.18.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
rich==13.7.1; python_full_version >= '3.7.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.30; python_version >= '3.7'
sqlalchemy-serializer==1.4.12
stack-data==0.6.3
tomli==2.0.1; python_version < '3.11' and python_version >= '3.7'
traitlets==5.14.3; python_version >= '3.8'
typing-extensions==4.11.0; python_version < '3.9'
wcwidth==0.2.13
werkzeug==3.0.3; python_version >= '3.8'
zipp==3.18.1; python_version >= '3.8'
